PRIS-4469: Accommodate multiple libs names when adding references to nuspec
Following the modification in the last commit, the creation of
Nuget packages in 1edit_amalgam failed because the reference
names for COM and WINRT reference groups now had additional
entries that named non-existent libraries. Here I have modified
the functionality that adds the reference to adopt the first
entry in the list, which happens to be the correct one (since
that was the original libs.names value).